requirements.txt
bcrypt==3.2.0
blinker==1.4
cffi==1.14.2
click==7.1.2
cryptography==3.1
Flask==1.1.2
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-SQLAlchemy==2.4.4
Flask-User==1.0.2.2
Flask-WTF==0.14.3
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
passlib==1.7.2
psycopg2==2.8.6
pycparser==2.20
six==1.15.0
SQLAlchemy==1.3.19
Werkzeug==1.0.1
WTForms==2.3.3

pandas~=1.1.2
Flask-Babel2~=0.1
alembic~=1.4.3
requests~=2.24.0
lxml~=4.5.2